All qualified versions to update
But still do not consider mutable SNAPSHOT versions. Most Kafka
versions are qualified, and we still want to update them.

If you're unhappy with some blood on the bleeding edge, you can still
prevent a dependency from being upgraded. But I think, for personal
projects, it's quite reasonable to ask for *-alpha and other *-dev
versions, provided that most of the time it's about adding new
functionalities that existing code does not target.